chore: refactor the directory to make things clean
