+++ src/admin.py
@@ -1,10 +1,10 @@
 from django.contrib import admin
 
-from .models import Customer, Subscription, Order, MenuItem, Payment
+from .models import customer, subscription, order, menu, payment
 
 # Register your models here.
-admin.site.register(Customer)
-admin.site.register(Subscription)
-admin.site.register(Order)
-admin.site.register(MenuItem)
-admin.site.register(Payment)
+admin.site.register(customer.Customer)
+admin.site.register(subscription.Subscription)
+admin.site.register(order.Order)
+admin.site.register(menu.MenuItem)
+admin.site.register(payment.Payment)
